Remove PCB files
Remove PCB files from Buck and Boost simulation projects since these
projects are intended to be pure simulation example and not to represent
any actual Buck or Boost implementation. When that is going to happen an
actual model of Buck or Boost driving stage and actual N-MOS model are
to be provided.